Move db stats functions to own file
